refactor: restful 진행
