+++ src/main/java/com/nhnacademy/bookstorefront/order/service/BookService.java
@@ -4,7 +4,7 @@ package com.nhnacademy.bookstorefront.order.service;
 import org.springframework.stereotype.Service;
 
 import com.nhnacademy.bookstorefront.order.dto.response.GetBookResponse;
-import com.nhnacademy.bookstorefront.order.feginclient.OrderServiceClient;
+import com.nhnacademy.bookstorefront.order.feignclient.OrderServiceClient;
 
 import lombok.RequiredArgsConstructor;
