+++ src/main/java/com/nhnacademy/bookstorefront/order/service/PaperTypeService.java
@@ -2,7 +2,7 @@ package com.nhnacademy.bookstorefront.order.service;
 
 import org.springframework.stereotype.Service;
 import com.nhnacademy.bookstorefront.order.dto.response.GetAllPaperResponse;
-import com.nhnacademy.bookstorefront.order.feginclient.OrderServiceClient;
+import com.nhnacademy.bookstorefront.order.feignclient.OrderServiceClient;
 
 import lombok.RequiredArgsConstructor;
